+++ cmd/gendoc/main.go
@@ -77,9 +77,9 @@ func main() {
 		fmt.Print(err)
 	}
 	for _, k := range keys {
-		_, err = mdf.WriteString(fmt.Sprintf(
+		_, err := fmt.Fprintf(mdf,
 			"## %s\n\n"+
-				"%s\n\n", k, md[k].Doc))
+				"%s\n\n", k, md[k].Doc)
 		if err != nil {
 			fmt.Print(err)
 		}
@@ -90,16 +90,16 @@ func main() {
 				size = len(e)
 			}
 			size /= len(md[k].Examples)
-			_, err := mdf.WriteString(fmt.Sprintf(
+			_, err := fmt.Fprintf(mdf,
 				"Output raw size: %d\n\n"+
-					"### Examples\n\n", size))
+					"### Examples\n\n", size)
 			if err != nil {
 				fmt.Print(err)
 			}
 			for n, e := range md[k].Examples {
-				_, err := mdf.WriteString(fmt.Sprintf(
+				_, err := fmt.Fprintf(mdf,
 					"#### %s\n"+
-						"```%s```\n\n", n, e))
+						"```%s```\n\n", n, e)
 				if err != nil {
 					fmt.Print(err)
 				}
@@ -264,23 +264,22 @@ func mustGetPackageName(astRoot string, f *ast.Package) string {
 	}
 	if firstKey == "" {
 		log.Fatalf("Package %q is composed of %d source files", f.Name, len(f.Files))
-	} else {
-		pkgAbs, err := filepath.Abs(filepath.Join(astRoot, filepath.Dir(firstKey)))
-		if err != nil {
-			log.Fatal(err)
-		}
-		goModPath, relPkgPath, err := findGoMod(pkgAbs)
-		if err != nil {
-			log.Fatal(err)
-		}
-		moduleName, err := getModuleNameFromGoMod(goModPath)
-		if err != nil {
-			log.Fatal(err)
-		}
-		importPath := filepath.Join(moduleName, relPkgPath)
-		return importPath
 	}
-	return ""
+
+	pkgAbs, err := filepath.Abs(filepath.Join(astRoot, filepath.Dir(firstKey)))
+	if err != nil {
+		log.Fatal(err)
+	}
+	goModPath, relPkgPath, err := findGoMod(pkgAbs)
+	if err != nil {
+		log.Fatal(err)
+	}
+	moduleName, err := getModuleNameFromGoMod(goModPath)
+	if err != nil {
+		log.Fatal(err)
+	}
+	importPath := filepath.Join(moduleName, relPkgPath)
+	return importPath
 }
 
 // execExampleMethod executes the method by starting go run and capturing the produced standard output
@@ -299,8 +298,7 @@ func execExampleMethod(methodFullPackage, methodPackage, methodName string) (str
 			fmt.Print(err)
 		}
 	}()
-
-	_, err = tf.WriteString(fmt.Sprintf(`package main
+	_, err = fmt.Fprintf(tf, `package main
 	import "%s"
 	import "fmt"
 
@@ -308,7 +306,7 @@ func execExampleMethod(methodFullPackage, methodPackage, methodName string) (str
 		str, _ := %s.%s()
 		fmt.Print(str)
 	}
-		`, methodFullPackage, methodPackage, methodName))
+		`, methodFullPackage, methodPackage, methodName)
 	if err != nil {
 		fmt.Print(err)
 	}